requirements.txt
anyio==3.7.1
asyncpg==0.29.0
certifi==2024.8.30
httpx==0.27.2
huggingface-hub==0.26.1
numpy==1.26.4
pandas==2.2.3
python-dotenv==1.0.0
Quart==0.19.9
regex==2024.9.11
requests==2.32.3
SQLAlchemy==2.0.21
